fix: 🐛 No such file or directory (os error 2)
Previously, when running the example shell scripts an error is thrown
"No such file or directory (os error 2)".

On dfx 0.6.26 the directory is expected to be `.dfx/local/canisters/reverse`,
for that reason a mkdir command creates the desired build directory.

The script now also skips build if the directory already already exists,
preceding the call to the canister reverse service.

On sucessful build, the wasm file is now moved instead of copied,
and artifacts are cleared from root.