Refactoring focused on invoice main.mo to use new libraries based utility module
This commit is focused on refactoring for src/invoice/main.mo to use new utilities module with some minor edits to ICPLedger.mo and Types.mo. Note that getDestinationAccountIdentifier and its associated type declarations were removed as it can now be created directly when creating an invoice. Additionally, during invoice creation the specified token type is checked before this address is created. Verify_invoice when used with the ICPLedger had two arguments added to its method argument type to limit the changes to invoice.mo which will be phased out before the merge of this branch is complete and are used so testing can demonstrating everything still works. Some minor revisions to match the code style guidelines was added to make it imperative flow control more clear.