refactored code base to make usability clearer in preparation for supporting more than icp
With new libraries utility module simplifying functionality and need for standardizing usability for integrating ICRC1, invoice type now uses a payment address that is of text type. Calls to invoice canister can now uniformly use human readable format for specifying transaction addresses, however a utility method for returning the blob form is still provided; in support of this, naming convention was updated to use 'address' as a suffix to indicate the human readable format. Note that all tests run successfully.