Use newest Internet Identity in encrypted notes dapps (#601)
Updates the encrypted-notes-dapp and encrypted-notes-dapp-vetkd to use the newest Internet Identity dev build.

This fixes the problem that only a single Internet Identity anchor can be created. In particular, attempting to create a second anchor led to the canister trapping with "the combination of timestamp 0 and nonce 0 has already been used". Especially for encrypted-notes-dapp-vetkd this was suboptimal because it prevented to show/prove that two different users cannot see each other's notes.

The dev build has some extra features that make Internet Identity easier to use in a development setting (such as a predictable captcha, no requirement for WebAuthn, etc.). Note that the dev-build was already used before, however.