fix: Update rust composite query to work again (#698)
This PR fixes the rust composite query to work again. Some of the changes:

Move dfx.json in the top level project directory for rust composite queries to not confuse dfx.
Add a Cargo.lock as per usual practice.
Use proper rust targets for the canisters.
Update readme and instructions a bit to build both the data partition and kv_frontend before installing the frontend and running queries against it.