Use default provisioned DFX version in encrypted-notes-vetkd dapps (#764)
Removes the custom DFX versions used for provisioning the Linux CI checks for the encrypted-notes-vetkd dapp for both Rust and Motoko. This is done because the custom DFX version, 0.14.2, is outdated. With this, the default DFX version defined in provision-linux.sh is used, which version is regularly updated to ensure the examples work with the latest version.