Use Wasm SIMD in the image classification example
The mainnet and dfx version 0.20.2-beta.0 support Wasm SIMD.
Sonos Tract also supports Wasm SIMD on `master`.

This PR updates the Sonos Tract dependency to the version that
supports Wasm SIMD.

Now image classification can run in a query. The example is
modified to add an option for running as a query or update.