feat: Add example for how to write unit-testable rust canisters (#1270)
This provides an example of best practices for structuring canisters for easier unit testing, as well as an example of a PocketIC test that automatically updates the wasm under test when running the test.  It contains multiple examples for how to write good tests in canisters that are easy to maintain.